+++ test/test_har_input.py
@@ -3,8 +3,7 @@
 import os
 
 from httpolice.inputs.har import har_input
-from httpolice.known import h, m
-from httpolice.structure import Unavailable, http2, http11
+from httpolice.structure import Unavailable
 
 
 def load_from_file(name):
@@ -12,133 +11,72 @@ def load_from_file(name):
     return list(har_input([path]))
 
 
-def test_http2bin_chrome():
-    exchanges = load_from_file('http2bin_chrome.har')
-
-    assert u'http2bin_chrome.har' in exchanges[0].request.remark
-    assert exchanges[0].request.version is None
-    assert u'http2bin_chrome.har' in exchanges[0].responses[0].remark
-    assert exchanges[0].responses[0].version is None
-    assert isinstance(exchanges[0].responses[0].body, Unavailable)
-
-    assert exchanges[1].request.target == u'https://http2bin.org/encoding/utf8'
-    assert not exchanges[1].responses[0].reason
-
-    assert exchanges[4].responses[0].body == b''
-
-    assert isinstance(exchanges[10].request.body, Unavailable)
-    assert isinstance(exchanges[10].request.decoded_body, Unavailable)
-    assert exchanges[10].request.unicode_body == (u'custname=qwedqwed&'
-                                                  u'custtel=dqwedwe&'
-                                                  u'custemail=&'
-                                                  u'size=medium&'
-                                                  u'delivery=&'
-                                                  u'comments=')
+def test_request_target_http11():
+    [exch] = load_from_file('chrome_text.har')
+    # Because this request has a Host header,
+    # we use the HTTP/1.x style of request target.
+    assert exch.request.target == u'/success.txt'
 
 
-def test_http2bin_firefox():
-    exchanges = load_from_file('http2bin_firefox.har')
-
-    assert exchanges[0].request.version == http2
-    assert exchanges[0].request.headers.connection.is_absent
-    assert isinstance(exchanges[0].responses[0].body, Unavailable)
-    assert isinstance(exchanges[0].responses[0].decoded_body, Unavailable)
-    assert exchanges[0].responses[0].unicode_body[:5] == u'{\n  "'
-    assert exchanges[0].responses[0].json_data['url'] == \
-        u'https://http2bin.org/get'
-
-    assert exchanges[5].responses[0].body == b''
-    assert exchanges[5].responses[0].decoded_body == b''
-    assert exchanges[5].responses[0].unicode_body == u''
-
-    assert isinstance(exchanges[7].responses[0].body, Unavailable)
-    assert len(exchanges[7].responses[0].decoded_body) == 1024
-
-    assert isinstance(exchanges[10].request.body, Unavailable)
-    assert isinstance(exchanges[10].request.decoded_body, Unavailable)
-    assert exchanges[10].request.unicode_body == (u'custname=ferferf&'
-                                                  u'custtel=rfwrefwerf&'
-                                                  u'custemail=&'
-                                                  u'size=medium&'
-                                                  u'delivery=&'
-                                                  u'comments=')
+def test_request_target_http2():
+    [exch] = load_from_file('chrome_http2.har')
+    # In HTTP/2 requests, Chrome doesn't send the Host header,
+    # so we use the full URL as the request target.
+    assert exch.request.target == u'https://vasiliy.faronov.name/'
 
 
-def test_spdy_chrome():
-    exchanges = load_from_file('spdy_chrome.har')
-    assert exchanges[0].request.version is None
-    assert exchanges[0].responses[0].version is None
-    assert exchanges[1].request.version is None
-    assert exchanges[1].responses[0].version is None
+def test_request_text():
+    [exch] = load_from_file('firefox_post_form.har')
+    assert u'firefox_post_form.har' in exch.request.remark
+    assert u'firefox_post_form.har' in exch.responses[0].remark
+    # Raw (unencoded) payload body is never available from HAR files.
+    assert isinstance(exch.request.body, Unavailable)
+    # Nor do we know the encoded payload as a bytestring.
+    assert isinstance(exch.request.decoded_body, Unavailable)
+    # Only the Unicode text of the body is present here.
+    assert exch.request.unicode_body.startswith(u'custname=Vasiliy')
 
 
-def test_spdy_firefox():
-    exchanges = load_from_file('spdy_firefox.har')
-    assert exchanges[0].responses[0].version is None
-    assert exchanges[1].responses[0].version is None
+def test_response_fail():
+    [exch] = load_from_file('chrome_https_fail.har')
+    assert exch.responses == []
 
 
-def test_xhr_chrome():
-    exchanges = load_from_file('xhr_chrome.har')
-    assert exchanges[0].request.target == u'/put'
-    assert exchanges[0].request.version == http11
-    assert exchanges[0].responses[0].version == http11
-    assert isinstance(exchanges[0].request.body, Unavailable)
-    assert isinstance(exchanges[0].request.decoded_body, Unavailable)
-    assert exchanges[0].request.unicode_body == u'wrfqerfqerferg45rfrqerf'
-    assert isinstance(exchanges[0].responses[0].body, Unavailable)
-    assert isinstance(exchanges[0].responses[0].decoded_body, Unavailable)
-    assert exchanges[0].responses[0].unicode_body[:5] == u'{\n  "'
-    assert exchanges[0].responses[0].json_data['data'] == \
-        u'wrfqerfqerferg45rfrqerf'
+def test_response_empty():
+    [exch] = load_from_file('firefox_empty.har')
+    assert exch.request.body == b''
+    assert exch.responses[0].body == b''
 
 
-def test_xhr_firefox():
-    exchanges = load_from_file('xhr_chrome.har')
-    assert exchanges[0].request.target == u'/put'
-    assert exchanges[0].request.version == http11
-    assert exchanges[0].responses[0].version == http11
+def test_response_text():
+    [exch] = load_from_file('chrome_text.har')
+    assert u'chrome_text.har' in exch.request.remark
+    assert u'chrome_text.har' in exch.responses[0].remark
+    # Raw (unencoded) payload body is never available from HAR files.
+    assert isinstance(exch.responses[0].body, Unavailable)
+    # Nor do we know the encoded payload as a bytestring.
+    assert isinstance(exch.responses[0].decoded_body, Unavailable)
+    # Only the Unicode text of the body is present here.
+    assert exch.responses[0].unicode_body == u'success\n'
 
 
-def test_httpbin_edge():
-    exchanges = load_from_file('httpbin_edge.har')
-
-    assert exchanges[0].request.target == u'/get'
-    assert exchanges[0].request.version is None
-    assert exchanges[0].responses[0].version is None
-    assert isinstance(exchanges[0].responses[0].body, Unavailable)
-    assert exchanges[0].responses[0].json_data['url'] == \
-        u'http://httpbin.org/get'
-
-    assert u'Unicode Demo' in exchanges[1].responses[0].unicode_body
-
-    assert exchanges[4].responses[0].body == b''
-    assert exchanges[5].responses[0].body == b''
-
-
-def test_xhr_edge():
-    exchanges = load_from_file('xhr_edge.har')
-    assert exchanges[1].request.method == m.DELETE
-    assert exchanges[1].request.body == b''
-    assert exchanges[2].request.method == u'FOO-BAR'
-
-
-def test_firefox_multiple_set_cookie():
-    exchanges = load_from_file('firefox_set_cookie.har')
-    [(_, _, (_, cookie1)), (_, _, (_, cookie2))] = \
-        exchanges[0].responses[0].headers.enumerate(h.set_cookie)
-    assert cookie1 == b'foo=bar'
-    assert cookie2 == b'baz=qux'
-
-
-def test_firefox_gif():
+def test_response_base64():
     exchanges = load_from_file('firefox_gif.har')
-    assert isinstance(exchanges[0].responses[0].body, Unavailable)
-    assert isinstance(exchanges[0].responses[0].decoded_body, Unavailable)
+    assert exchanges[-2].responses[0].decoded_body.startswith(b'GIF89')
+
+
+def test_response_bad_base64():
+    [exch] = load_from_file('bad_base64.har')
+    # Can't do anything with a broken base64 encoding.
+    assert isinstance(exch.responses[0].body, Unavailable)
+    assert isinstance(exch.responses[0].decoded_body, Unavailable)
+    assert isinstance(exch.responses[0].unicode_body, Unavailable)
 
 
 def test_fiddler_connect():
     exchanges = load_from_file('fiddler+ie11_connect.har')
+    assert u'fiddler+ie11_connect.har' in exchanges[0].request.remark
+    assert u'fiddler+ie11_connect.har' in exchanges[0].responses[0].remark
     assert exchanges[0].request.target == u'httpbin.org:443'
     assert exchanges[0].request.body == b''
     assert exchanges[0].request.decoded_body == b''
@@ -146,3 +84,11 @@ def test_fiddler_connect():
     assert exchanges[0].responses[0].headers[u'EndTime'].is_absent
     assert exchanges[0].responses[0].body == b''
     assert exchanges[0].responses[0].decoded_body == b''
+
+
+def test_firefox_304():
+    # Firefox includes content on 304 responses, but we ignore it.
+    [exch, _] = load_from_file('firefox_304.har')
+    assert exch.responses[0].body == b''
+    assert exch.responses[0].decoded_body == b''
+    assert exch.responses[0].unicode_body == u''